+++ common/src/main/java/com/lowdragmc/lowdraglib/jei/ModularUIRecipeCategory.java
@@ -1,26 +1,28 @@
 package com.lowdragmc.lowdraglib.jei;
 
 import com.lowdragmc.lowdraglib.gui.ingredient.IRecipeIngredientSlot;
+import com.lowdragmc.lowdraglib.gui.widget.DraggableScrollableWidgetGroup;
 import com.lowdragmc.lowdraglib.gui.widget.Widget;
+import com.lowdragmc.lowdraglib.utils.Position;
 import mezz.jei.api.gui.builder.IRecipeLayoutBuilder;
 import mezz.jei.api.gui.builder.IRecipeSlotBuilder;
-import mezz.jei.api.ingredients.IIngredientType;
+import mezz.jei.api.gui.builder.ITooltipBuilder;
+import mezz.jei.api.gui.ingredient.IRecipeSlotsView;
+import mezz.jei.api.gui.widgets.IRecipeExtrasBuilder;
 import mezz.jei.api.recipe.IFocusGroup;
 import mezz.jei.api.recipe.RecipeIngredientRole;
 import mezz.jei.api.recipe.category.IRecipeCategory;
-import mezz.jei.api.runtime.IClickableIngredient;
 import net.minecraft.MethodsReturnNonnullByDefault;
+import net.minecraft.client.Minecraft;
+import net.minecraft.client.gui.GuiGraphics;
 
 import javax.annotation.Nullable;
 import javax.annotation.ParametersAreNonnullByDefault;
-import java.util.ArrayList;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
 
 /**
  * @author KilaBash
- * @date: 2022/04/30
+ * @date 2022/04/30
  * @implNote ModularUIRecipeCategory
  */
 @MethodsReturnNonnullByDefault
@@ -29,32 +31,73 @@ public abstract class ModularUIRecipeCategory<T extends ModularWrapper<?>> imple
 
     @Override
     public void setRecipe(IRecipeLayoutBuilder builder, T wrapper, IFocusGroup focuses) {
+        wrapper.setRecipeWidget(0, 0);
+
         List<Widget> flatVisibleWidgetCollection = wrapper.modularUI.getFlatWidgetCollection();
-        for (int i = 0; i < flatVisibleWidgetCollection.size(); i++) {
-            Widget widget = flatVisibleWidgetCollection.get(i);
+        for (Widget widget : flatVisibleWidgetCollection) {
             if (widget instanceof IRecipeIngredientSlot slot) {
                 var role = mapToRole(slot.getIngredientIO());
+                if (widget.getParent() instanceof DraggableScrollableWidgetGroup) {
+                    // don't add the JEI widget at all if we have a draggable group, let the draggable widget handle it instead.
+                    if (role == null) {
+                        builder.addInvisibleIngredients(RecipeIngredientRole.INPUT).addIngredientsUnsafe(slot.getXEIIngredients());
+                        builder.addInvisibleIngredients(RecipeIngredientRole.OUTPUT).addIngredientsUnsafe(slot.getXEIIngredients());
+                    } else {
+                        builder.addInvisibleIngredients(role).addIngredientsUnsafe(slot.getXEIIngredients());
+                    }
+                    continue;
+                }
+                Position pos = widget.getPosition();
+                IRecipeSlotBuilder slotBuilder;
                 if (role == null) { // both
-                    addJEISlot(builder, slot, RecipeIngredientRole.INPUT, i);
-                    addJEISlot(builder, slot, RecipeIngredientRole.OUTPUT, i);
+                    addJEISlot(builder, slot, RecipeIngredientRole.INPUT, pos.x, pos.y);
+                    slotBuilder = addJEISlot(builder, slot, RecipeIngredientRole.OUTPUT, pos.x, pos.y);
                 } else {
-                    addJEISlot(builder, slot, role, i);
+                    slotBuilder = addJEISlot(builder, slot, role, pos.x, pos.y);
+                }
+                int width = widget.getSize().width;
+                int height = widget.getSize().height;
+
+                slotBuilder.setBackground(IGui2IDrawable.toDrawable(widget.getBackgroundTexture(), width, height), -1, -1);
+                slotBuilder.setOverlay(IGui2IDrawable.toDrawable(widget.getOverlay(), width, height), -1, -1);
+                widget.setActive(false);
+                widget.setVisible(false);
+                if (slot instanceof com.lowdragmc.lowdraglib.gui.widget.SlotWidget slotW) {
+                    slotW.setDrawHoverOverlay(false).setDrawHoverTips(false);
+                } else if (slot instanceof com.lowdragmc.lowdraglib.gui.widget.TankWidget tankW) {
+                    tankW.setDrawHoverOverlay(false).setDrawHoverTips(false);
+                    slotBuilder.setFluidRenderer(1, false, width - 2, height - 2);
                 }
             }
         }
-
     }
 
-    private static void addJEISlot(IRecipeLayoutBuilder builder, IRecipeIngredientSlot slot, RecipeIngredientRole role, int index) {
-        IRecipeSlotBuilder slotBuilder = builder.addSlot(role, index, -1);
-        //noinspection rawtypes
-        Map<IIngredientType, List> map = new HashMap<>();
-        //noinspection unchecked
-        slot.getXEIIngredients().stream()
-                .filter(IClickableIngredient.class::isInstance)
-                .map(IClickableIngredient.class::cast)
-                .forEach(clickableIngredient -> map.computeIfAbsent(clickableIngredient.getTypedIngredient().getType(), a -> new ArrayList<>()).add(clickableIngredient.getTypedIngredient().getIngredient()));
-        map.forEach(slotBuilder::addIngredients);
+    @Override
+    public void createRecipeExtras(IRecipeExtrasBuilder builder, T wrapper, IFocusGroup focuses) {
+        builder.addGuiEventListener(new ModularUIGuiEventListener<>(wrapper));
+        builder.addWidget(new ModularForegroundRecipeWidget(wrapper));
+    }
+
+    @Override
+    public void draw(T recipe, IRecipeSlotsView recipeSlotsView, GuiGraphics guiGraphics, double mouseX, double mouseY) {
+        recipe.draw(guiGraphics, (int) mouseX, (int) mouseY, Minecraft.getInstance().getFrameTime());
+    }
+
+    @Override
+    public void getTooltip(ITooltipBuilder tooltip, T recipe, IRecipeSlotsView recipeSlotsView, double mouseX, double mouseY) {
+        IRecipeCategory.super.getTooltip(tooltip, recipe, recipeSlotsView, mouseX, mouseY);
+        if (recipe.tooltipTexts != null && !recipe.tooltipTexts.isEmpty()) {
+            tooltip.addAll(recipe.tooltipTexts);
+        }
+        if (recipe.tooltipComponent != null) {
+            tooltip.add(recipe.tooltipComponent);
+        }
+    }
+
+    private static IRecipeSlotBuilder addJEISlot(IRecipeLayoutBuilder builder, IRecipeIngredientSlot slot, RecipeIngredientRole role, int x, int y) {
+        IRecipeSlotBuilder slotBuilder = builder.addSlot(role, x, y);
+        slotBuilder.addIngredientsUnsafe(slot.getXEIIngredients());
+        return slotBuilder;
     }
 
     @Nullable