+++ forge/src/main/java/com/lowdragmc/lowdraglib/forge/CommonProxyImpl.java
@@ -10,11 +10,13 @@ import com.lowdragmc.lowdraglib.client.renderer.block.forge.RendererBlockEntityI
 import com.lowdragmc.lowdraglib.plugin.ILDLibPlugin;
 import com.lowdragmc.lowdraglib.plugin.LDLibPlugin;
 import com.lowdragmc.lowdraglib.syncdata.TypedPayloadRegistries;
+import com.lowdragmc.lowdraglib.test.NoRendererTestBlock;
 import com.lowdragmc.lowdraglib.test.TestBlock;
 import com.lowdragmc.lowdraglib.test.TestBlockEntity;
 import com.lowdragmc.lowdraglib.test.TestItem;
 import com.lowdragmc.lowdraglib.test.forge.TestBlockEntityImpl;
 import com.lowdragmc.lowdraglib.utils.ReflectionUtils;
+import net.minecraft.world.item.BlockItem;
 import net.minecraft.world.item.Item;
 import net.minecraft.world.level.block.Block;
 import net.minecraft.world.level.block.entity.BlockEntityType;
@@ -37,6 +39,8 @@ public class CommonProxyImpl {
         if (Platform.isDevEnv()) {
             BLOCKS.register("test", () -> TestBlock.BLOCK);
             ITEMS.register("test", () -> TestItem.ITEM);
+            BLOCKS.register("test_2", () -> NoRendererTestBlock.BLOCK);
+            ITEMS.register("test_2", () -> new BlockItem(NoRendererTestBlock.BLOCK, new Item.Properties()));
             TestBlockEntityImpl.TYPE = BLOCK_ENTITY_TYPES.register("test", () -> BlockEntityType.Builder.of(TestBlockEntity::new, TestBlock.BLOCK).build(null));
         }