+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 # ChangeLog
 
-* fix TrackedDummyWorld leaking client level instances of ~300mb
+* fix draggable widgets not adding their contents to JEI pages' lookups
+* fix graphics getting overlayed on JEI slot textures